fix err

+++ src/hal/hal_flash.h
@@ -9,5 +9,5 @@
 
 
 uint8_t hal_flash_erase(void);
-void hal_flash_write(uint32_t addr ,uint32_t *buff, uint32_t num);
+void hal_flash_write(uint32_t addr ,uint16_t *buff, uint32_t num);
 #endif
